fix the menulist (make it play nice with periodic refreshing).
before the addition of the timer for periodic matches updates
the menulist was supposed to be refreshed manually by the user,
and we didn't take into account the fact that every time you
clear it and refill it loses its navigational parameters, i.e.
which entry was selected and what's the offset if there are
more entries than the visible window (Maxvisitems).

so now we store those and reconfigure the menulist, adapting
to its new state, so that the user experiences as little
disruption as possible.